Webinar20/src/test/java/sampleShop/homePage/HomePageTest.java
package sampleShop.homePage;

import org.openqa.selenium.By;
import org.testng.Assert;
import org.testng.annotations.Test;
import sampleShop.BaseTest;
import pl.testuj.sampleshop.pages.HomePage;

public class HomePageTest extends BaseTest {

    @Test
    public void checkCategoriesInMenu() {
        HomePage homePage = new HomePage(driver);
        Assert.assertTrue(homePage.checkMenuItemExists("Clothes"));
        Assert.assertTrue(homePage.checkMenuItemExists("Produkty powiązane"));
        Assert.assertTrue(homePage.checkMenuItemExists("Art"));

//        Assert.assertFalse(homePage.checkMenuItemExists("Kontakt z nami"));
    }

//    @Test
//    public void fakeTest() {
//        driver.findElement(By.className("fakeClassName"));
//    }
}
